[ENH] Move FullyConnectedModule to layers/_mlp/; credit v1 author in DecoderMLP_v2
