+++ pytorch_forecasting/layers/_mlp/__init__.py
@@ -1,0 +1,7 @@
+"""
+Fully connected (MLP) layers.
+"""
+
+from pytorch_forecasting.layers._mlp._fully_connected import FullyConnectedModule
+
+__all__ = ["FullyConnectedModule"]
